bugfix 2894681: wrong string for reputation.
I changed the way some strings are accessed - each record in the Strings file should be accessed exactly once, by the lowest possible type. I should find a better way to assure this.

I don't want to remove the Strings file compleatly, because it's a hook for internationalization (is that how it's spelled?). I'll shop around for a good solution for this problem evantually.